feat: extract shared init logic into @internals/shared
Move PluginOption type, availablePlugins, pluginDefaultConfigs,
KUBB_CONFIG_FILENAME, initDefaults, and generateConfigFile() from
packages/cli and packages/mcp into a new private @internals/shared
package so both consumers share a single source of truth.
