+++ internals/shared/src/index.ts
@@ -1,0 +1,3 @@
+export type { PluginOption } from './types.ts'
+export { KUBB_CONFIG_FILENAME, initDefaults, availablePlugins, pluginDefaultConfigs } from './constants.ts'
+export { generateConfigFile } from './init.ts'
