+++ internals/shared/src/init.ts
@@ -1,0 +1,31 @@
+import { pluginDefaultConfigs } from './constants.ts'
+import type { PluginOption } from './types.ts'
+
+export function generateConfigFile(selectedPlugins: PluginOption[], inputPath: string, outputPath: string): string {
+  const imports = selectedPlugins.map((plugin) => `import { ${plugin.importName} } from '${plugin.packageName}'`).join('\n')
+
+  const pluginConfigs = selectedPlugins
+    .map((plugin) => {
+      const config = (pluginDefaultConfigs as Record<string, string>)[plugin.value] ?? `${plugin.importName}()`
+      return `    ${config},`
+    })
+    .join('\n')
+
+  return `import { defineConfig } from 'kubb'
+${imports}
+
+export default defineConfig({
+  root: '.',
+  input: {
+    path: '${inputPath}',
+  },
+  output: {
+    path: '${outputPath}',
+    clean: true,
+  },
+  plugins: [
+${pluginConfigs}
+  ],
+})
+`
+}
