+++ internals/shared/src/types.ts
@@ -1,0 +1,8 @@
+export type PluginOption = {
+  value: string
+  label: string
+  hint?: string
+  packageName: string
+  importName: string
+  category: 'types' | 'client' | 'framework' | 'validation' | 'testing' | 'mocks' | 'documentation' | 'ai'
+}
