+++ packages/cli/package.json
@@ -71,6 +71,7 @@
     "tinyexec": "catalog:"
   },
   "devDependencies": {
+    "@internals/shared": "workspace:*",
     "@internals/utils": "workspace:*",
     "@kubb/adapter-oas": "workspace:*"
   },
