+++ packages/cli/src/runners/init.ts
@@ -5,120 +5,8 @@ import { styleText } from 'node:util'
 import * as clack from '@clack/prompts'
 import type { PackageManagerInfo } from '@internals/utils'
 import { detectPackageManager } from '@internals/utils'
-import { initDefaults, KUBB_CONFIG_FILENAME, pluginDefaultConfigs } from '../constants.ts'
+import { availablePlugins, generateConfigFile, initDefaults, KUBB_CONFIG_FILENAME, type PluginOption } from '@internals/shared'
 import { hasPackageJson, initPackageJson, installPackages } from '../utils/packageManager.ts'
-
-type PluginOption = {
-  value: string
-  label: string
-  hint?: string
-  packageName: string
-  importName: string
-  category: 'types' | 'client' | 'framework' | 'validation' | 'testing' | 'mocks' | 'documentation' | 'ai'
-}
-
-const availablePlugins: PluginOption[] = [
-  {
-    value: 'plugin-ts',
-    label: 'TypeScript',
-    hint: 'Recommended',
-    packageName: '@kubb/plugin-ts',
-    importName: 'pluginTs',
-    category: 'types',
-  },
-  {
-    value: 'plugin-client',
-    label: 'Client (Fetch/Axios)',
-    packageName: '@kubb/plugin-client',
-    importName: 'pluginClient',
-    category: 'client',
-  },
-  {
-    value: 'plugin-react-query',
-    label: 'React Query / TanStack Query',
-    packageName: '@kubb/plugin-react-query',
-    importName: 'pluginReactQuery',
-    category: 'framework',
-  },
-  {
-    value: 'plugin-vue-query',
-    label: 'Vue Query',
-    packageName: '@kubb/plugin-vue-query',
-    importName: 'pluginVueQuery',
-    category: 'framework',
-  },
-  {
-    value: 'plugin-zod',
-    label: 'Zod Schemas',
-    packageName: '@kubb/plugin-zod',
-    importName: 'pluginZod',
-    category: 'validation',
-  },
-  {
-    value: 'plugin-faker',
-    label: 'Faker.js Mocks',
-    packageName: '@kubb/plugin-faker',
-    importName: 'pluginFaker',
-    category: 'mocks',
-  },
-  {
-    value: 'plugin-msw',
-    label: 'MSW Handlers',
-    packageName: '@kubb/plugin-msw',
-    importName: 'pluginMsw',
-    category: 'mocks',
-  },
-  {
-    value: 'plugin-cypress',
-    label: 'Cypress Tests',
-    packageName: '@kubb/plugin-cypress',
-    importName: 'pluginCypress',
-    category: 'testing',
-  },
-  {
-    value: 'plugin-mcp',
-    label: 'MCP Server (AI / Model Context Protocol)',
-    packageName: '@kubb/plugin-mcp',
-    importName: 'pluginMcp',
-    category: 'ai',
-  },
-  {
-    value: 'plugin-redoc',
-    label: 'ReDoc Documentation',
-    packageName: '@kubb/plugin-redoc',
-    importName: 'pluginRedoc',
-    category: 'documentation',
-  },
-]
-
-function generateConfigFile(selectedPlugins: PluginOption[], inputPath: string, outputPath: string): string {
-  const imports = selectedPlugins.map((plugin) => `import { ${plugin.importName} } from '${plugin.packageName}'`).join('\n')
-
-  const pluginConfigs = selectedPlugins
-    .map((plugin) => {
-      const config = (pluginDefaultConfigs as Record<string, string>)[plugin.value] ?? `${plugin.importName}()`
-      return `    ${config},`
-    })
-    .join('\n')
-
-  return `import { defineConfig } from 'kubb'
-${imports}
-
-export default defineConfig({
-  root: '.',
-  input: {
-    path: '${inputPath}',
-  },
-  output: {
-    path: '${outputPath}',
-    clean: true,
-  },
-  plugins: [
-${pluginConfigs}
-  ],
-})
-`
-}
 
 function cancelAndExit(message = 'Operation cancelled.'): never {
   clack.cancel(message)
