+++ tsconfig.json
@@ -8,6 +8,7 @@
       "#mocks": ["./internals/utils/src/mocks.ts"],
       "@kubb/core/mocks": ["./packages/core/src/mocks.ts"],
       "@internals/utils": ["./internals/utils/src/index.ts"],
+      "@internals/shared": ["./internals/shared/src/index.ts"],
       "@kubb/ast": ["./packages/ast/src/index.ts"],
       "@kubb/core": ["./packages/core/src/index.ts"],
       "@kubb/agent": ["./packages/agent/src/index.ts"],
